Add JUnit test for Formatter class
