--- src/test/java/FormatterTest.java
@@ -1,16 +1,0 @@
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class FormatterTest {
-    @Test
-    public void dummyTest() {
-        assertEquals(2, 2);
-    }
-
-    @Test
-    public void formatResponseTest() {
-        String response = "hi";
-        assertEquals(2, 2);
-    }
-}
